+++ src/e_regex.h
@@ -104,9 +104,9 @@ typedef struct {
 } RxMatchRes;
 
 RxNode *RxCompile(const char *Regexp);
-int RxExecMatch(RxNode *Regexp, const char *Data, int Len, const char *Start, RxMatchRes *Match, unsigned int RxOpt = RX_CASE);
-int RxExec(RxNode *Regexp, const char *Data, int Len, const char *Start, RxMatchRes *Match, unsigned int RxOpt = RX_CASE);
-int RxReplace(const char *rep, const char *Src, int len, RxMatchRes match, char **Dest, int *Dlen);
+int RxExecMatch(RxNode *Regexp, const char *Data, size_t Len, const char *Start, RxMatchRes *Match, unsigned int RxOpt = RX_CASE);
+int RxExec(RxNode *Regexp, const char *Data, size_t Len, const char *Start, RxMatchRes *Match, unsigned int RxOpt = RX_CASE);
+int RxReplace(const char *rep, const char *Src, size_t len, RxMatchRes match, char **Dest, int *Dlen);
 void RxFree(RxNode *Node);
 
 #endif
